feat: implement hero data fetching and update BannerAboutUs to use new hook
