+++ src/shared/api/getHero.ts
@@ -1,6 +1,6 @@
 import { collection, getDocs } from 'firebase/firestore'
-import { db } from '../../../services/firebase'
-import type { HeroItem } from '../sections/hero/type'
+import { db } from '../../services/firebase'
+import type { HeroItem } from '../../slices/home/sections/hero/type'
 
 export async function getHeroItems(): Promise<HeroItem[]> {
   const snapshot = await getDocs(collection(db, 'hero'))
