+++ src/slices/home/sections/hero/Hero.tsx
@@ -1,12 +1,12 @@
 import { Skeleton } from '@/shared/components/Skeleton'
-import { useHero } from '../../hooks/useHero'
+import { useHero } from '../../../../shared/hooks/useHero'
 import HeroBackground from './components/HeroBackground'
 import HeroContentDesktop from './components/HeroContentDesktop'
 import HeroContentMobile from './components/HeroContentMobile'
 import type { HeroItem } from './type'
 
 export default function Hero() {
-  const { items, loading, error } = useHero()
+  const { items, loading, error } = useHero("")
 
   if (loading) {
     return (
@@ -33,8 +33,8 @@ export default function Hero() {
   if (error) return <div className="py-16 text-center text-red-600">{error}</div>
   if (!items.length) return <div className="py-16 text-center">Sin contenido.</div>
 
-  const left: HeroItem = items[0]
-  const right: HeroItem = items[1] || items[0]
+  const left: HeroItem = items[1]
+  const right: HeroItem = items[2] || items[1]
 
   return (
     <section className="relative text-white">
